cleanup/requirements.txt
# MinerU Cleanup Service Requirements
# 清理服务的依赖包

# Environment variables
python-dotenv>=1.0.0

# Task scheduling
schedule>=1.2.0

# S3 Storage Support (Optional, for S3 storage mode)
s3fs>=2024.3.0

